requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
plotly==5.19.0
pyyaml==6.0.1
ccxt==4.2.15
ta-lib==0.4.28
scikit-learn==1.4.1
tensorflow==2.15.0
python-telegram-bot==20.8
psutil==5.9.8
requests==2.31.0
python-dotenv==1.0.1 